Added support for GIFs on the client. Code is kinda fugly, but I'll clean it up later.

+++ android_client/app/src/main/java/com/zigzag/client_app/model/AnimationSetImageData.java
@@ -4,15 +4,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class AnimationSetImageData extends ImageData {
-    private final double timeBetweenFrames;
+    private final long timeBetweenFrames;
     private final List<TileData> framesDesc;
 
-    public AnimationSetImageData(double timeBetweenFrames, List<TileData> framesDesc) {
+    public AnimationSetImageData(long timeBetweenFrames, List<TileData> framesDesc) {
         this.timeBetweenFrames = timeBetweenFrames;
         this.framesDesc = framesDesc;
     }
 
-    public double getTimeBetweenFrames() {
+    public long getTimeBetweenFrames() {
         return timeBetweenFrames;
     }
